Added package.json so 'npm install' can be used to install dependencies
Also updated the readme to use full Markdown while I was in there messing with installation instructions.

Finally, updated the script to use the 'date' field for box office gross since that changed in the new revision of NPMJS-rt-scraper. This is all included in the same commit mostly because of package-lock.json (readme installation instructions depend on package.json, package-log.json depends on package.json, MMM-RottenTomatoes.js depends on package-lock.json).